Altered SCSS, changed some tags in HTML - Added figcaptions to all category svgs. Wasn't sure whether to use <p> or <figcaption>, but thought figcaption made the most sense.
- Resized all images to make it easier to position
them.

- Adjusted the CSS and HTML to allow both
svgs and figcaptions to function as a singular
element, in order to create a blue
underline/effect when hovering over.

- Wrapped logo svg and figcaption in a <div>
element.
- Retitled CSS accordingly.
- Removed "sticky" tag as it had no effect.
- Changed finn-logo width to 40%.

- Text in header is uneven, trying to fix it.
- Set tag "align-items: center;" to nav ul li.
- Changed nav ul li svg padding to 1px.

- Added border-bottom to all header items.
- Added display:flex, flex-direction:column and
align-items:left to nav ul li.
It sort of worked, but now the text is too high
up.
- The border underline lines up with the hr well
enough, and I don't want to add more clutter to
the code.